mood_music_project/music/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'folders', views.MoodFolderViewSet, basename='folder')
router.register(r'songs', views.SongViewSet, basename='song')

urlpatterns = [
    path('auth/register/', views.register, name='register'),
    path('auth/login/', views.login, name='login'),
    path('auth/logout/', views.logout, name='logout'),
    path('', include(router.urls)),
]